// apps/front-end/src/http/types/CreateCouponBodyDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreateCouponBodyDto = {
  /**
   * @type string
   */
  productId: string;
  /**
   * @minLength 0
   * @maxLength 100
   * @type number
   */
  discountPercentage: number;
  /**
   * @maxLength 9007199254740991
   * @type integer
   */
  maxQuantity: number;
};