// apps/front-end/src/http/types/GetCouponResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GetCouponResponseDto = {
  /**
   * @type object
   */
  coupon: {
    /**
     * @type string
     */
    id: string;
    /**
     * @type number
     */
    currentQuantity: number;
    /**
     * @type number
     */
    discountPercentage: number;
    /**
     * @type number
     */
    maxQuantity: number;
    /**
     * @type string
     */
    productId: string;
    /**
     * @type string
     */
    userId: string;
    /**
     * @type string
     */
    createdAt?: string | null;
    /**
     * @type string
     */
    updatedAt?: string | null;
  };
};