// apps/front-end/src/http/types/UpdateCouponBodyDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateCouponBodyDto = {
  /**
   * @minLength 0
   * @maxLength 100
   * @type number | undefined
   */
  discountPercentage?: number;
  /**
   * @maxLength 9007199254740991
   * @type integer | undefined
   */
  maxQuantity?: number;
  /**
   * @minLength 0
   * @maxLength 9007199254740991
   * @type integer | undefined
   */
  currentQuantity?: number;
};